Split Job functionality into different file
